Some UI / UX improvements

+++ lib/screens/my_workouts/my_workouts_cubit.dart
@@ -10,25 +10,27 @@ part 'my_workouts_cubit.freezed.dart';
 class MyWorkoutsViewState with _$MyWorkoutsViewState {
   const factory MyWorkoutsViewState({
     @Default([]) List<Workout> workouts,
-    @Default(false) bool isLoading
+    @Default(false) bool isLoading,
   }) = _MyWorkoutsViewState;
 }
 
 abstract class MyWorkoutsNavigator {
   Future<void> showErrorMessage(String message);
 
+  /// Returns true if the user has confirmed the action
+  Future<bool> showConfirmationMessage(
+      {required String title, required String message});
+
   /// Returns true if the workouts need to be reloaded when returning
   Future<bool> goToWorkoutDetails(Workout workout);
 
   /// Returns true if the workouts need to be reloaded when returning
   Future<bool> goToAddWorkout();
-
-  /// Returns true if the workouts need to be reloaded when returning
-  Future<bool> goToEditWorkout(Workout workout);
 }
 
 class MyWorkoutsCubit extends Cubit<MyWorkoutsViewState> {
-  MyWorkoutsCubit(this._repository, this._navigator) : super(const MyWorkoutsViewState()) {
+  MyWorkoutsCubit(this._repository, this._navigator)
+      : super(const MyWorkoutsViewState()) {
     _loadWorkouts();
   }
 
@@ -38,7 +40,7 @@ class MyWorkoutsCubit extends Cubit<MyWorkoutsViewState> {
   Future<void> _loadWorkouts() async {
     emit(state.copyWith(isLoading: true));
     final workoutsResult = await _repository.getWorkouts();
-    switch(workoutsResult) {
+    switch (workoutsResult) {
       case Failure(error: final error):
         emit(state.copyWith(isLoading: false));
         await _navigator.showErrorMessage(error.toString());
@@ -63,14 +65,14 @@ class MyWorkoutsCubit extends Cubit<MyWorkoutsViewState> {
     }
   }
 
-  Future<void> onEditWorkout(Workout workout) async {
-    final shouldReload = await _navigator.goToEditWorkout(workout);
-    if (shouldReload) {
-      _loadWorkouts();
-    }
-  }
-
   Future<void> onDeleteWorkout(Workout workout) async {
+    final confirmed = await _navigator.showConfirmationMessage(
+      title: 'Confirmation',
+      message: 'Delete workout ${workout.name}?',
+    );
+    if (!confirmed) {
+      return;
+    }
     final result = await _repository.deleteWorkout(workout.id);
     if (result is Failure) {
       await _navigator.showErrorMessage(result.error.toString());
